Require Vectors to have the same dimension
Vectors should only act on other Vectors of the same dimension. This
required, for the moment, to remove itertools, and use lists for
scaling.

    >>> it.repeat(scalar)  # abused map taking only smallest needed
                           # however if len used, would be inf
    >>> it.repeat(scalar, len(self))  # preferred, but has no len()
                                      # don't want to wrap in tuple
    >>> [scalar] * len(self)  # has len(), does the job

A custom Exception is raised, based off ArithmeticError. This seem
appropriate.

To have access to the length function, scalar_product was changed from
a staticmethod to a classmethod. Rather than explicitly wrapping it,
changed map -> cls._map.