Fix _same_length to correctly test all iterables
Previous implementation folded the lengths of the iterables

    >>> ((len(i1) == len(i2)) == len(i3)) == len(i4)

However this would always evaluate as False if the number of iterables
were greater than 2, due to:

    >>> True == len(i3)  # always False

Explicitly taking the first length, and generating comparisons to that
length, is cleaner. The built-in `all` correctly reduces the results
to a single boolean, even if it is a single iterable given to the
_same_length function:

    >>> all([])  # True

It is clearer what is going on, and has the correct functionality.